requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1
django-crispy-forms==2.3
djangorestframework==3.14.0
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
